pyproject.toml
[tool.poetry]
name = "echobot"
version = "0.1.0"
description = "A Chatbot for QQ Group Based on Nonebot 2"
authors = ["jks15satoshi <jks15satoshi@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"
fastapi_utils = "^0.2.1"
typing_extensions = "^3.7.4"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
